requirements.txt
fastapi
uvicorn
langchain
langchain-community
sentence-transformers
faiss-cpu
pydantic
ollama
python-dotenv
